Use tracetest instead of oteltest in test mod
