+++ instrumentation/github.com/labstack/echo/otelecho/echo_test.go
@@ -26,123 +26,13 @@ import (
 	"github.com/labstack/echo/v4"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/oteltest"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 
 	b3prop "go.opentelemetry.io/contrib/propagators/b3"
-	"go.opentelemetry.io/otel/attribute"
-	oteltrace "go.opentelemetry.io/otel/trace"
 )
-
-func TestChildSpanFromGlobalTracer(t *testing.T) {
-	otel.SetTracerProvider(oteltest.NewTracerProvider())
-
-	router := echo.New()
-	router.Use(Middleware("foobar"))
-	router.GET("/user/:id", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
-		_, ok := span.(*oteltest.Span)
-		assert.True(t, ok)
-		return c.NoContent(200)
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	router.ServeHTTP(w, r)
-	assert.Equal(t, http.StatusOK, w.Result().StatusCode, "should call the 'user' handler")
-}
-
-func TestChildSpanFromCustomTracer(t *testing.T) {
-	provider := oteltest.NewTracerProvider()
-
-	router := echo.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-	router.GET("/user/:id", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
-		_, ok := span.(*oteltest.Span)
-		assert.True(t, ok)
-		return c.NoContent(200)
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	router.ServeHTTP(w, r)
-	assert.Equal(t, http.StatusOK, w.Result().StatusCode, "should call the 'user' handler")
-}
-
-func TestTrace200(t *testing.T) {
-	sr := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
-
-	router := echo.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-	router.GET("/user/:id", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
-		mspan, ok := span.(*oteltest.Span)
-		require.True(t, ok)
-		assert.Equal(t, attribute.StringValue("foobar"), mspan.Attributes()["http.server_name"])
-		id := c.Param("id")
-		return c.String(200, id)
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	// do and verify the request
-	router.ServeHTTP(w, r)
-	response := w.Result()
-	require.Equal(t, http.StatusOK, response.StatusCode)
-
-	// verify traces look good
-	spans := sr.Completed()
-	require.Len(t, spans, 1)
-	span := spans[0]
-	assert.Equal(t, "/user/:id", span.Name())
-	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
-	assert.Equal(t, attribute.StringValue("foobar"), span.Attributes()["http.server_name"])
-	assert.Equal(t, attribute.IntValue(http.StatusOK), span.Attributes()["http.status_code"])
-	assert.Equal(t, attribute.StringValue("GET"), span.Attributes()["http.method"])
-	assert.Equal(t, attribute.StringValue("/user/123"), span.Attributes()["http.target"])
-	assert.Equal(t, attribute.StringValue("/user/:id"), span.Attributes()["http.route"])
-}
-
-func TestError(t *testing.T) {
-	sr := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
-
-	// setup
-	router := echo.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-	wantErr := errors.New("oh no")
-	// configure a handler that returns an error and 5xx status
-	// code
-	router.GET("/server_err", func(c echo.Context) error {
-		return wantErr
-	})
-	r := httptest.NewRequest("GET", "/server_err", nil)
-	w := httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	response := w.Result()
-	assert.Equal(t, http.StatusInternalServerError, response.StatusCode)
-
-	// verify the errors and status are correct
-	spans := sr.Completed()
-	require.Len(t, spans, 1)
-	span := spans[0]
-	assert.Equal(t, "/server_err", span.Name())
-	assert.Equal(t, attribute.StringValue("foobar"), span.Attributes()["http.server_name"])
-	assert.Equal(t, attribute.IntValue(http.StatusInternalServerError), span.Attributes()["http.status_code"])
-	assert.Equal(t, attribute.StringValue("oh no"), span.Attributes()["echo.error"])
-	// server errors set the status
-	assert.Equal(t, codes.Error, span.StatusCode())
-}
 
 func TestErrorOnlyHandledOnce(t *testing.T) {
 	router := echo.New()
@@ -165,7 +55,7 @@ func TestGetSpanNotInstrumented(t *testing.T) {
 	router := echo.New()
 	router.GET("/ping", func(c echo.Context) error {
 		// Assert we don't have a span on the context.
-		span := oteltrace.SpanFromContext(c.Request().Context())
+		span := trace.SpanFromContext(c.Request().Context())
 		ok := !span.SpanContext().IsValid()
 		assert.True(t, ok)
 		return c.String(200, "ok")
@@ -196,7 +86,7 @@ func TestPropagationWithGlobalPropagators(t *testing.T) {
 	router := echo.New()
 	router.Use(Middleware("foobar", WithTracerProvider(provider)))
 	router.GET("/user/:id", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
+		span := trace.SpanFromContext(c.Request().Context())
 		assert.Equal(t, sc.TraceID(), span.SpanContext().TraceID())
 		assert.Equal(t, sc.SpanID(), span.SpanContext().SpanID())
 		return c.NoContent(200)
@@ -227,7 +117,7 @@ func TestPropagationWithCustomPropagators(t *testing.T) {
 	router := echo.New()
 	router.Use(Middleware("foobar", WithTracerProvider(provider), WithPropagators(b3)))
 	router.GET("/user/:id", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
+		span := trace.SpanFromContext(c.Request().Context())
 		assert.Equal(t, sc.TraceID(), span.SpanContext().TraceID())
 		assert.Equal(t, sc.SpanID(), span.SpanContext().SpanID())
 		return c.NoContent(200)
@@ -248,7 +138,7 @@ func TestSkipper(t *testing.T) {
 	router := echo.New()
 	router.Use(Middleware("foobar", WithSkipper(skipper)))
 	router.GET("/ping", func(c echo.Context) error {
-		span := oteltrace.SpanFromContext(c.Request().Context())
+		span := trace.SpanFromContext(c.Request().Context())
 		assert.False(t, span.SpanContext().HasSpanID())
 		assert.False(t, span.SpanContext().HasTraceID())
 		return c.NoContent(200)
