+++ instrumentation/github.com/labstack/echo/otelecho/test/echo_test.go
@@ -1,0 +1,140 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Based on https://github.com/DataDog/dd-trace-go/blob/8fb554ff7cf694267f9077ae35e27ce4689ed8b6/contrib/gin-gonic/gin/gintrace_test.go
+
+package test
+
+import (
+	"errors"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/labstack/echo/v4"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"go.opentelemetry.io/contrib/instrumentation/github.com/labstack/echo/otelecho"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
+
+	"go.opentelemetry.io/otel/attribute"
+	oteltrace "go.opentelemetry.io/otel/trace"
+)
+
+func TestChildSpanFromGlobalTracer(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
+	otel.SetTracerProvider(provider)
+
+	router := echo.New()
+	router.Use(otelecho.Middleware("foobar"))
+	router.GET("/user/:id", func(c echo.Context) error {
+		return c.NoContent(200)
+	})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	router.ServeHTTP(w, r)
+	assert.Equal(t, http.StatusOK, w.Result().StatusCode, "should call the 'user' handler")
+	assert.Len(t, sr.Ended(), 1)
+}
+
+func TestChildSpanFromCustomTracer(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
+
+	router := echo.New()
+	router.Use(otelecho.Middleware("foobar", otelecho.WithTracerProvider(provider)))
+	router.GET("/user/:id", func(c echo.Context) error {
+		return c.NoContent(200)
+	})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	router.ServeHTTP(w, r)
+	assert.Equal(t, http.StatusOK, w.Result().StatusCode, "should call the 'user' handler")
+	assert.Len(t, sr.Ended(), 1)
+}
+
+func TestTrace200(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
+
+	router := echo.New()
+	router.Use(otelecho.Middleware("foobar", otelecho.WithTracerProvider(provider)))
+	router.GET("/user/:id", func(c echo.Context) error {
+		id := c.Param("id")
+		return c.String(200, id)
+	})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	// do and verify the request
+	router.ServeHTTP(w, r)
+	response := w.Result()
+	require.Equal(t, http.StatusOK, response.StatusCode)
+
+	// verify traces look good
+	spans := sr.Ended()
+	require.Len(t, spans, 1)
+	span := spans[0]
+	assert.Equal(t, "/user/:id", span.Name())
+	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
+	attrs := span.Attributes()
+	assert.Contains(t, attrs, attribute.String("http.server_name", "foobar"))
+	assert.Contains(t, attrs, attribute.Int("http.status_code", http.StatusOK))
+	assert.Contains(t, attrs, attribute.String("http.method", "GET"))
+	assert.Contains(t, attrs, attribute.String("http.target", "/user/123"))
+	assert.Contains(t, attrs, attribute.String("http.route", "/user/:id"))
+}
+
+func TestError(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := trace.NewTracerProvider(trace.WithSpanProcessor(sr))
+
+	// setup
+	router := echo.New()
+	router.Use(otelecho.Middleware("foobar", otelecho.WithTracerProvider(provider)))
+	wantErr := errors.New("oh no")
+	// configure a handler that returns an error and 5xx status
+	// code
+	router.GET("/server_err", func(c echo.Context) error {
+		return wantErr
+	})
+	r := httptest.NewRequest("GET", "/server_err", nil)
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, r)
+	response := w.Result()
+	assert.Equal(t, http.StatusInternalServerError, response.StatusCode)
+
+	// verify the errors and status are correct
+	spans := sr.Ended()
+	require.Len(t, spans, 1)
+	span := spans[0]
+	assert.Equal(t, "/server_err", span.Name())
+	attrs := span.Attributes()
+	assert.Contains(t, attrs, attribute.String("http.server_name", "foobar"))
+	assert.Contains(t, attrs, attribute.Int("http.status_code", http.StatusInternalServerError))
+	assert.Contains(t, attrs, attribute.String("echo.error", "oh no"))
+	// server errors set the status
+	assert.Equal(t, codes.Error, span.Status().Code)
+}
